Renamed dl_util.h to dl_internal_util.h to avoid include-issues
Since I stupidly created a file that has a name which another file has, with the wrong include/path setups you get conflicting includes. This changelist fixes that.
