+++ src/dl_internal_util.h
